requirements.txt
openai>=1.5.0
pydantic>=2.0.0
pandas>=1.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0